Revert Named Colors change
The previous change caused a startup-regression with Color because it spends time building up a dictionary as soon as the Color type is touched (in its static ctor). Changing this code back to use pattern-matching, which forces us to allocate a string, for now. Once we have a compiler that supports pattern-matching on a ReadOnlySpan<char>, we can do this allocation free.

This is no worse than the original implementation, which also always allocated the lowered string to do pattern matching on.